+++ cch/src/main/java/wft/hahn/neo4j/cch/model/VertexPaths.java
@@ -63,7 +63,7 @@ public class VertexPaths {
         }
 
         @Override
-        public float weight() {
+        public int weight() {
             return 0;
         }
     }
